Used doubles instead of floats for better accuracy.

+++ DNN_/N_Network_/HyperParams/HyperParams.cpp
@@ -27,22 +27,22 @@ HyperParams::HyperParams() {
 }
 
 //To allocate memory at compile time
-//std::array<float, HIDDEN_NODES> HyperParams::allocate_memory(std::array<Node*, HIDDEN_NODES> hidden_nodes,
+//std::array<double, HIDDEN_NODES> HyperParams::allocate_memory(std::array<Node*, HIDDEN_NODES> hidden_nodes,
 //                                                             std::array<Node*, HIDDEN_NODES> next_nodes) {}
 //
 //returns the computed activations
 //might be too abstracted
-std::vector<float> HyperParams::hidden_unit(std::vector<float> activations) {
+std::vector<double> HyperParams::hidden_unit(std::vector<double> activations) {
     return node.computed_activations;
 }
 
 //This is designed to iterate over h_params.hidden_nodes, and for each outer iteration
 // an innner iteration is performed to connect the nodes of that current layer to the 
 // nodes of the next layer.
-std::vector<std::vector<float>> HyperParams::hidden_layer(std::array<Node*, HIDDEN_NODES> &hidden_nodes,
+std::vector<std::vector<double>> HyperParams::hidden_layer(std::array<Node*, HIDDEN_NODES> &hidden_nodes,
                                                           std::array<Node*, HIDDEN_NODES> &next_nodes) {
     //initializing hidden layer matrix
-    std::vector<std::vector<float>> activations(2, std::vector<float>(HIDDEN_NODES, 0.0f));
+    std::vector<std::vector<double>> activations(2, std::vector<double>(HIDDEN_NODES, 0.0f));
 
     for (int i = 0; i < hidden_nodes.size(); i++) {
         if (hidden_nodes[i]) {
@@ -57,9 +57,9 @@ std::vector<std::vector<float>> HyperParams::hidden_layer(std::array<Node*, HIDD
 }
 
 //overloaded function using vectors
-std::vector<std::vector<float>> HyperParams::hidden_layer(std::vector<Node*> &hidden_nodes,
+std::vector<std::vector<double>> HyperParams::hidden_layer(std::vector<Node*> &hidden_nodes,
                                                           std::vector<Node*> &next_nodes) {
-    std::vector<std::vector<float>> activations(2, std::vector<float>(HIDDEN_NODES, 0.0f));
+    std::vector<std::vector<double>> activations(2, std::vector<double>(HIDDEN_NODES, 0.0f));
     for (int i = 0; i < hidden_nodes.size(); i++) {
         if (hidden_nodes[i]) {
             for (int j = 0; j < next_nodes.size(); j++) {
@@ -78,7 +78,7 @@ std::vector<HyperParams::Node> HyperParams::deep_net_constructor(std::vector<
                                                                  HIDDEN_NODES>> &layers) {
     std::vector<Node> layer_activations;
     for (int i = 0; i < HIDDEN_LAYERS; i++) {
-        std::vector<std::vector<float>> layer_activations = hidden_layer(layers[i], 
+        std::vector<std::vector<double>> layer_activations = hidden_layer(layers[i], 
                                                                          layers[i+1]);
     }
     return layer_activations;
@@ -88,7 +88,7 @@ std::vector<HyperParams::Node> HyperParams::input_layer_constructor(std::vector<
                                                                     std::vector<Node*> &layers) {
     std::vector<Node> layer_activations;
     for (int i = 0; i < INPUT_NODES; i++) {
-        std::vector<std::vector<float>> layer_activations = hidden_layer(input_nodes, 
+        std::vector<std::vector<double>> layer_activations = hidden_layer(input_nodes, 
                                                                          layers);
     }
     return layer_activations;
@@ -98,14 +98,14 @@ std::vector<HyperParams::Node> HyperParams::output_layer_constructor(std::vector
                                                                      std::vector<Node*> &output_nodes) {
     std::vector<Node> layer_activations;
     for (int i = 0; i < HIDDEN_LAYERS; i++) {
-        std::vector<std::vector<float>> layer_activations = hidden_layer(layers, 
+        std::vector<std::vector<double>> layer_activations = hidden_layer(layers, 
                                                                          output_nodes);
     }
     return layer_activations;
 }
 
 //change the learning rate
-void HyperParams::change_learning_rate(float &new_rate) {
+void HyperParams::change_learning_rate(double &new_rate) {
     node.learning_rate =  new_rate;
 }
